requirements.txt
yfinance
pandas
matplotlib
ta
pyyaml